move credential to connection
